refactor: 	1. create group to manage basic concept test project
