+++ CGConceptTest/src/Learn_OpenGL/gladConf/premake5.lua
@@ -1,0 +1,41 @@
+project "GLADConf"
+    kind "ConsoleApp"
+    language "C++"
+    cppdialect "C++17"
+
+    targetdir ("bin/" .. outputdir .. "/%{prj.name}")
+    objdir ("bin-int/" .. outputdir .. "/%{prj.name}")
+    
+    files
+    {
+        "**.h",
+        "**.cpp"
+    }
+
+    includedirs
+    {
+        "%{IncludeDir.GLFW}",
+        "%{IncludeDir.GLAD}"        
+    }
+
+    links
+    {
+        "GLFW",
+        "GLAD",
+        "opengl32.lib"
+    }
+
+    filter "system:windows"
+        systemversion "latest"
+
+	filter "configurations:Debug"
+        systemversion "latest"
+		runtime "Debug"
+        symbols "on"
+        staticruntime "off"
+
+	filter "configurations:Release"
+        systemversion "latest"
+		runtime "Release"
+        optimize "on"
+        staticruntime "On"
