+++ CGHackerTest/src/SandBox.cpp
@@ -1,91 +1,2 @@
 #include <iostream>
 #include "cg.h"
-#include <glad/glad.h>
-#include <GLFW/glfw3.h>
-
-void framebuffer_size_callback(GLFWwindow* window, int width, int height)
-{
-	glViewport(0, 0, width, height);
-}
-
-void processInput(GLFWwindow* window)
-{
-	if (glfwGetKey(window, GLFW_KEY_ESCAPE) == GLFW_PRESS)
-	{
-		glfwSetWindowShouldClose(window, true);
-	}
-}
-
-int test_glfw()
-{
-	GLFWwindow* window;
-
-	if (!glfwInit())
-		return -1;
-
-	window = glfwCreateWindow(640, 480, "Hello World", NULL, NULL);
-	if (!window)
-	{
-		glfwTerminate();
-		return -1;
-	}
-
-	glfwMakeContextCurrent(window);
-	while (!glfwWindowShouldClose(window))
-	{
-		glClear(GL_COLOR_BUFFER_BIT);
-
-		glfwSwapBuffers(window);
-
-		glfwPollEvents();
-	}
-	
-	glfwTerminate();
-}
-int test_glad()
-{
-	glfwInit();
-
-	// Create window use glfw
-
-	GLFWwindow* window = glfwCreateWindow(800, 600, "Glad Test", NULL, NULL);
-	if (window == NULL)
-	{
-		std::cout << "Failed to create GLFW window" << std::endl;
-		glfwTerminate();
-		return -1;
-	}
-
-	// Create opengl context
-
-	glfwMakeContextCurrent(window);
-	glfwSetFramebufferSizeCallback(window, framebuffer_size_callback);
-
-	if (!gladLoadGLLoader((GLADloadproc)glfwGetProcAddress))
-	{
-		std::cout << "Failed to initalize GLAD" << std::endl;
-		return -1;
-	}
-
-
-	while (!glfwWindowShouldClose(window))
-	{
-		processInput(window);
-
-		glClearColor(0.2f, 0.3f, 0.3f, 1.0f);
-		glClear(GL_COLOR_BUFFER_BIT);
-
-		glfwSwapBuffers(window);
-		glfwPollEvents();
-	}
-
-	glfwTerminate();
-	return 0;
-}
-
-int main()
-{
-	//test_glfw();
-	test_glad();
-	std::cin.get();
-}
