+++ premakes/learn_opengl.lua
@@ -1,0 +1,4 @@
+group "CGConceptTest/LearnOpenGL"
+    include "CGConceptTest/src/Learn_OpenGL/glfwConf"
+    include "CGConceptTest/src/Learn_OpenGL/gladConf"
+group ""
